Trait hrpc_build::Service

pub trait Service {
    type Comment: AsRef<str>;
    type Method: Method;

    const CODEC_PATH: &'static str;

    fn name(&self) -> &str;

    fn package(&self) -> &str;

    fn identifier(&self) -> &str;

    fn methods(&self) -> &[Self::Method];

    fn comment(&self) -> &[Self::Comment];
}

Service generation trait.

This trait can be implemented and consumed by client::generate and server::generate to allow any codegen module to generate service abstractions.

Associated Types

type Comment: AsRef<str>

Comment type.

type Method: Method

Method type.

Loading content...
Associated Constants

const CODEC_PATH: &'static str

Path to the codec.

Loading content...
Required methods

fn name(&self) -> &str

Name of service.

fn package(&self) -> &str

Package name of service.

fn identifier(&self) -> &str

Identifier used to generate type name.

fn methods(&self) -> &[Self::Method]

Methods provided by service.

fn comment(&self) -> &[Self::Comment]

Get comments about this item.

Loading content...